pyproject.toml
[project]
name = "metachat"
version = "0.1.0"
description = "MetaChat is an innovative messaging platform that leverages AI-driven meta functionalities to deliver seamless, context-aware interactions."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aipolabs>=0.0.1b7",
    "Flask>=2.2.2",
    "twilio>=7.17.0",
    "python-dotenv>=0.21.0"
]

[dependency-groups]
dev = [
    "anthropic>=0.49.0",
    "crewai>=0.108.0",
    "langchain>=0.3.21",
    "langchain-openai>=0.3.11",
    "llama-index>=0.12.26",
    "openai>=1.68.2",
    "rich>=13.9.4",
    "ruff>=0.11.2",
    "cloudinary"
]
